Remove `query_consensus_states` from the `ChainEndpoint` trait
