+++ crates/relayer/src/chain/cosmos.rs
@@ -64,6 +64,7 @@ use crate::chain::cosmos::fee::maybe_register_counterparty_payee;
 use crate::chain::cosmos::gas::{calculate_fee, mul_ceil};
 use crate::chain::cosmos::query::account::get_or_fetch_account;
 use crate::chain::cosmos::query::balance::{query_all_balances, query_balance};
+use crate::chain::cosmos::query::consensus_state::query_consensus_state_heights;
 use crate::chain::cosmos::query::custom::cross_chain_query_via_rpc;
 use crate::chain::cosmos::query::denom_trace::query_denom_trace;
 use crate::chain::cosmos::query::status::query_status;
@@ -82,7 +83,7 @@ use crate::chain::requests::*;
 use crate::chain::tracking::TrackedMsgs;
 use crate::client_state::{AnyClientState, IdentifiedAnyClientState};
 use crate::config::{parse_gas_prices, ChainConfig, GasPrice};
-use crate::consensus_state::{AnyConsensusState, AnyConsensusStateWithHeight};
+use crate::consensus_state::AnyConsensusState;
 use crate::denom::DenomTrace;
 use crate::error::Error;
 use crate::event::monitor::{EventMonitor, TxMonitorCmd};
@@ -92,8 +93,7 @@ use crate::light_client::tendermint::LightClient as TmLightClient;
 use crate::light_client::{LightClient, Verified};
 use crate::misbehaviour::MisbehaviourEvidence;
 use crate::util::pretty::{
-    PrettyConsensusStateWithHeight, PrettyHeight, PrettyIdentifiedChannel,
-    PrettyIdentifiedClientState, PrettyIdentifiedConnection,
+    PrettyIdentifiedChannel, PrettyIdentifiedClientState, PrettyIdentifiedConnection,
 };
 
 pub mod batch;
@@ -1080,104 +1080,11 @@ impl ChainEndpoint for CosmosSdkChain {
         &self,
         request: QueryConsensusStateHeightsRequest,
     ) -> Result<Vec<ICSHeight>, Error> {
-        crate::time!("query_consensus_state_heights");
-        crate::telemetry!(query, self.id(), "query_consensus_state_heights");
-
-        // Helper function to diagnose if the QueryConsensusStateHeightsRequest is unsupported
-        // by matching on the error details.
-        fn is_unsupported(status: &tonic::Status) -> bool {
-            if status.code() != tonic::Code::Unimplemented {
-                return false;
-            }
-
-            status
-                .message()
-                .contains("unknown method ConsensusStateHeights")
-        }
-
-        let mut client = self
-            .block_on(
-                ibc_proto::ibc::core::client::v1::query_client::QueryClient::connect(
-                    self.grpc_addr.clone(),
-                ),
-            )
-            .map_err(Error::grpc_transport)?;
-
-        let grpc_request = tonic::Request::new(request.clone().into());
-        let grpc_response = self.block_on(client.consensus_state_heights(grpc_request));
-
-        if let Err(ref e) = grpc_response {
-            if is_unsupported(e) {
-                warn!("QueryConsensusStateHeights is not supported by the chain, falling back on QueryConsensusStates");
-
-                let states = self.query_consensus_states(QueryConsensusStatesRequest {
-                    client_id: request.client_id,
-                    pagination: request.pagination,
-                })?;
-
-                return Ok(states.into_iter().map(|cs| cs.height).collect());
-            }
-        }
-
-        let response = grpc_response.map_err(Error::grpc_status)?.into_inner();
-
-        let heights: Vec<_> = response
-            .consensus_state_heights
-            .into_iter()
-            .filter_map(|h| {
-                ICSHeight::try_from(h.clone())
-                    .map_err(|e| {
-                        warn!(
-                            "failed to parse consensus state height {}. Error: {}",
-                            PrettyHeight(&h),
-                            e
-                        )
-                    })
-                    .ok()
-            })
-            .collect();
-
-        Ok(heights)
-    }
-
-    fn query_consensus_states(
-        &self,
-        request: QueryConsensusStatesRequest,
-    ) -> Result<Vec<AnyConsensusStateWithHeight>, Error> {
-        crate::time!("query_consensus_states");
-        crate::telemetry!(query, self.id(), "query_consensus_states");
-
-        let mut client = self
-            .block_on(
-                ibc_proto::ibc::core::client::v1::query_client::QueryClient::connect(
-                    self.grpc_addr.clone(),
-                ),
-            )
-            .map_err(Error::grpc_transport)?;
-
-        let request = tonic::Request::new(request.into());
-        let response = self
-            .block_on(client.consensus_states(request))
-            .map_err(Error::grpc_status)?
-            .into_inner();
-
-        let consensus_states: Vec<_> = response
-            .consensus_states
-            .into_iter()
-            .filter_map(|cs| {
-                AnyConsensusStateWithHeight::try_from(cs.clone())
-                    .map_err(|e| {
-                        warn!(
-                            "failed to parse consensus state {}. Error: {}",
-                            PrettyConsensusStateWithHeight(&cs),
-                            e
-                        )
-                    })
-                    .ok()
-            })
-            .collect();
-
-        Ok(consensus_states)
+        self.block_on(query_consensus_state_heights(
+            self.id(),
+            &self.grpc_addr,
+            request,
+        ))
     }
 
     fn query_consensus_state(
